refactor: extract boilerplate
